+++ shared.webpack.config.js
@@ -3,9 +3,7 @@
 
 'use strict';
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
 const merge = require('merge-options');
-// eslint-disable-next-line @typescript-eslint/no-var-requires
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
 
 /**