+++ server/src/common/language.ts
@@ -162,19 +162,36 @@ export const booleanFunctions: readonly string[] = [
 /**
  * ChoiceScript built-in variables
  */
-export const builtinVariables: readonly string[] = [
-	"choice_subscribe_allowed", "choice_register_allowed", "choice_registered",
-	"choice_is_web", "choice_is_steam", "choice_is_steam_deck", "choice_is_ios_app",
-	"choice_is_ipad_app", "choice_is_android_app", "choice_is_omnibus_app",
-	"choice_is_amazon_app",
-	"choice_is_advertising_supported", "choice_is_trial", "choice_release_date",
-	"choice_prerelease", "choice_kindle", "choice_randomtest", "choice_quicktest",
-	"choice_linenum", "choice_scene",
-	"choice_restore_purchases_allowed", "choice_save_allowed",
-	"choice_time_stamp", "choice_nightmode", "choice_title",
-	"choice_purchase_supported", "choice_purchased_adfree",
-	"choice_saved_checkpoint", "choice_just_restored_checkpoint"
+const choiceVariableAppNames = [ "ios", "ipad", "android", "omnibus", "amazon" ];
+const choiceVariableIsNames = [
+	"web",
+	"steam(_deck)?",
+	"(" + choiceVariableAppNames.join("|") + ")_app",
+	"advertising_supported",
+	"trial"
 ];
+const choiceVariableNames = [
+	"(subscribe|register|restore_purchases|save)_allowed",
+	"registered",
+	"is_(" + choiceVariableIsNames.join("|") + ")",
+	"release_date",
+	"prerelease",
+	"kindle",
+	"(random|quick)test",
+	"linenum",
+	"scene",
+	"time_stamp",
+	"nightmode",
+	"title",
+	"saved_checkpoint(_[a-zA-Z0-9_]+)?",
+	"reuse",
+	"user_restored",
+	"restore_name",
+	"just_restored_checkpoint",
+	"purchase_supported",
+	"purchased_adfree"
+];
+export const builtinVariables = new RegExp(`^${choiceVariableNames.join("|")}$`);
 
 /**
  * ChoiceScript param variables